+++ prj/vs2022/CvtOnnx.vcxproj.filters
@@ -114,6 +114,9 @@
     <ClCompile Include="..\..\src\Cvt\OnnxRuntime\ConvertExpNode.cpp">
       <Filter>OnnxRuntime\E</Filter>
     </ClCompile>
+    <ClCompile Include="..\..\src\Cvt\OnnxRuntime\ConvertExpandNode.cpp">
+      <Filter>OnnxRuntime\E</Filter>
+    </ClCompile>
     <ClCompile Include="..\..\src\Cvt\OnnxRuntime\ConvertFlattenNode.cpp">
       <Filter>OnnxRuntime\F</Filter>
     </ClCompile>
